pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-coach"
version = "0.1.0"
description = "Chess coaching tool: engine analysis + LLM natural language explanations"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "python-chess>=1.10",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "click>=8.1",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "hypothesis>=6.0",
    "mypy>=1.8",
    "ruff>=0.3",
    "tox>=4.0",
]

[tool.setuptools.package-data]
"chess_coach" = ["web/static/**/*", "web/static/*"]

[project.scripts]
chess-coach = "chess_coach.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
